+++ backend/utils/extract_quantitative_table.py
@@ -1,5 +1,8 @@
 import pdfplumber
 
+from suppress_stderr import suppress_stderr
+from extract_tables_from_pages import extract_tables_from_pages
+
 
 def is_quantitative_table_start(page, header_keywords):
     table = page.extract_table()
@@ -23,7 +26,7 @@ def is_quantitative_table_start(page, header_keywords):
 def is_table_end(page):
     text = page.extract_text()
     if not text:
-        return True  # עמוד ריק מסיים טבלה
+        return True
 
     end_text = ".עצומה בצמה טירשתב ןיבו תינכתה תוארוהב ןיב ,תורחא תויללכ תוארוה לע ,הריתס לש הרקמב ,רבוג וז הלבטב רומאה"
 
@@ -35,26 +38,26 @@ def extract_quantitative_table(pdf_path, header_keywords):
     table_pages = []
 
     table_started = False
-    with pdfplumber.open(pdf_path) as pdf:
-        for page_num, page in enumerate(pdf.pages):
+    with suppress_stderr():
+        with pdfplumber.open(pdf_path) as pdf:
+            for page_num, page in enumerate(pdf.pages):
 
-            if not table_started:
-                if is_quantitative_table_start(page, header_keywords):
-                    table_pages.append(page_num + 1)
-                    table_started = True
-                else:
-                    continue  # continue to the next page
+                if not table_started:
+                    if is_quantitative_table_start(page, header_keywords):
+                        table_pages.append(page_num + 1)
+                        table_started = True
+                    else:
+                        continue  # continue to the next page
 
-            if table_started:
-                if is_table_end(page):
-                    table_started = False
-                    table_pages.append(page_num + 1)
-                    break  # סוף הטבלה
-    return table_pages
+                if table_started:
+                    if is_table_end(page):
+                        table_started = False
+                        table_pages.append(page_num + 1)
+                        break  # סוף הטבלה
+        return table_pages
 
 
 if __name__ == "__main__":
-    # pdf_path = r"C:\Users\dpere\Documents\python-projects\web-gis-scraper-desktop-electron\backend\101-0925198_תדפיס הוראות התכנית_חתום לאישור_1.pdf"
     pdf_path = r"C:\Users\dpere\Documents\python-projects\web-gis-scraper-desktop-electron\backend\101-0135004_תדפיס הוראות התכנית_חתום לאישור_1.pdf"
 
     header_keywords = [
@@ -62,6 +65,6 @@ if __name__ == "__main__":
         "שרגמ לדוג",
         "הינב יחטש",
         'ד"חי רפסמ',
-    ]  # מילות מפתח רלוונטיות לטבלאות כמותיות
-    table = extract_quantitative_table(pdf_path, header_keywords)
-    print(table)
+    ]
+    pages = extract_quantitative_table(pdf_path, header_keywords)
+    extract_tables_from_pages(pdf_path, pages)
