+++ backend/utils/suppress_stderr.py
@@ -1,0 +1,14 @@
+import contextlib
+import os
+import sys
+
+
+@contextlib.contextmanager
+def suppress_stderr():
+    with open(os.devnull, "w") as devnull:
+        old_stderr = sys.stderr
+        sys.stderr = devnull
+        try:
+            yield
+        finally:
+            sys.stderr = old_stderr
